♻️ 서버 디렉토리 구조 전면 수정
