+++ .gitignore
@@ -8,9 +8,8 @@
 .env.test.local
 .env.production.local
 .env
-.firebase
 
 npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
-firebase-debug.log*
+
